Add CustomButton, CustomTextfield, and AuthScreen Components
Created new files for `CustomButton` and `CustomTextfield` widgets in the `lib/common/widgets` directory. These widgets are designed to be reusable components for buttons and text fields throughout the application.

Implemented the `AuthScreen` in the `lib/features/auth/screens` directory, which includes a stateful widget for handling user authentication. The screen allows users to switch between sign-up and sign-in modes, and it uses the newly created `CustomButton` and `CustomTextfield` widgets for form inputs.

Updated the `main.dart` file to include the new `AuthScreen` as the home screen and configured the app's theme to use custom colors from `GlobalVariables`.

Added a `router.dart` file to handle routing within the application, with a default route that displays a "Screen does not exist!" message for undefined routes.