Add Account Services and Fetch Orders Functionality
Created a new `AccountServices` class in the `account_services.dart` file to handle fetching user orders from the backend. Implemented the `fetchMyOrders` method to retrieve orders and parse them into a list of `Order` objects. Updated the `Orders` widget to use the fetched orders instead of a temporary list, and added an `initState` method to call the `fetchMyOrders` function on widget initialization. Also, added a new `Order` model to represent order data, including products, quantities, and status. This commit completes the integration of order fetching and display in the account section.